# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Api Swag App
# during dev, should point to your local machine
host: 127.0.0.1:7500
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - multipart/form-data
# format of the responses to the client (Accepts)
produces:
  - application/json
  - text/html
securityDefinitions:
  OAuth2:
    type: oauth2
    flow: application
    tokenUrl: http://localhost:7500/accesstoken
    scopes:
      scope1: Description here
      scope2: Description here
      scope3: Description here
      scope4: Description here
...
paths:
  /verifyToken:
    x-swagger-router-controller: login
    get:
      description: Verify access token
      # used as the method name of the controller
      operationId: verifyToken
      parameters:
        - name: Authorization
          in: header
          description: Access Token to Verify
          required: true
          type: string
        - name: scopes
          in: query
          description: Scopes to verify
          type: string
      responses:
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /refreshToken:
    x-swagger-router-controller: login
    post:
      description: Refresh token
      # used as the method name of the controller
      operationId: refreshToken
      parameters:
        - name: refreshToken
          in: body
          description: Access Token to Verify
          required: true
          schema:
            required:
              - refresh_token
            properties:
              refresh_token:
                type: string
      responses:
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...